+++ lib/plug/router/utils.ex
@@ -67,6 +67,23 @@ defmodule Plug.Router.Utils do
   end
 
   @doc """
+  Builds a list of path param names and var match pairs that can bind
+  to dynamic path segment values. Excludes params with underscores;
+  otherwise, the compiler will warn about used underscored variables
+  when they are unquoted in the macro.
+
+  ## Examples
+
+      iex> Plug.Router.Utils.build_path_params_match([:id])
+      [{"id", {:id, [], nil}}]
+  """
+  def build_path_params_match(vars) do
+    vars
+    |> Enum.map(&{Atom.to_string(&1), Macro.var(&1, nil)})
+    |> Enum.reject(&match?({"_" <> _var, _macro}, &1))
+  end
+
+  @doc """
   Forwards requests to another Plug at a new path.
   """
   def forward(%Plug.Conn{path_info: path, script_name: script} = conn, new_path, target, opts) do